Renamed forEach*
