+++ src/forEach.test.mjs
@@ -3,6 +3,7 @@ import { expect, test } from '@jest/globals'
 import forEach from './forEach.mjs'
 import Deferred from './Deferred.mjs'
 import { range } from 'itertools'
+import delay from './delay.mjs'
 
 test('forEach base', async () => {
   const arr = [...range(6)]
@@ -10,7 +11,7 @@ test('forEach base', async () => {
   arr.forEach((v) => { called[v] = 0 })
   await forEach(arr, async (x) => {
     called[x] += 1
-  })
+  }, 2)
   expect(called[0]).toBe(1)
   expect(called[1]).toBe(1)
   expect(called[2]).toBe(1)
@@ -25,7 +26,7 @@ test('forEach no async', async () => {
   arr.forEach((v) => { called[v] = 0 })
   await forEach(arr, (x) => {
     called[x] += 1
-  })
+  }, 2)
   expect(called[0]).toBe(1)
   expect(called[1]).toBe(1)
   expect(called[2]).toBe(1)
@@ -44,7 +45,72 @@ test('forEach concurrency', async () => {
     called[x] += 1
     ds[x].resolve()
     await d.promise
-  })
+  }, 2)
+  await ds[1].promise
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(0)
+  expect(called[3]).toBe(0)
+  expect(called[4]).toBe(0)
+  expect(called[5]).toBe(0)
+  await delay()
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(0)
+  expect(called[3]).toBe(0)
+  expect(called[4]).toBe(0)
+  expect(called[5]).toBe(0)
+  d.resolve()
+  await p
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(1)
+  expect(called[3]).toBe(1)
+  expect(called[4]).toBe(1)
+  expect(called[5]).toBe(1)
+})
+
+test('forEach infinite concurrency base', async () => {
+  const arr = [...range(6)]
+  const called = {}
+  arr.forEach((v) => { called[v] = 0 })
+  await forEach(arr, async (x) => {
+    called[x] += 1
+  }, Number.POSITIVE_INFINITY)
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(1)
+  expect(called[3]).toBe(1)
+  expect(called[4]).toBe(1)
+  expect(called[5]).toBe(1)
+})
+
+test('forEach infinite concurrency no async', async () => {
+  const arr = [...range(6)]
+  const called = {}
+  arr.forEach((v) => { called[v] = 0 })
+  await forEach(arr, (x) => {
+    called[x] += 1
+  }, Number.POSITIVE_INFINITY)
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(1)
+  expect(called[3]).toBe(1)
+  expect(called[4]).toBe(1)
+  expect(called[5]).toBe(1)
+})
+
+test('forEach infinite concurrency concurrency', async () => {
+  const arr = [...range(6)]
+  const called = {}
+  arr.forEach((v) => { called[v] = 0 })
+  const d = new Deferred()
+  const ds = arr.map(() => new Deferred())
+  const p = forEach(arr, async (x) => {
+    called[x] += 1
+    ds[x].resolve()
+    await d.promise
+  }, Number.POSITIVE_INFINITY)
   await ds[5].promise
   expect(called[0]).toBe(1)
   expect(called[1]).toBe(1)
@@ -52,6 +118,64 @@ test('forEach concurrency', async () => {
   expect(called[3]).toBe(1)
   expect(called[4]).toBe(1)
   expect(called[5]).toBe(1)
+  d.resolve()
+  await p
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(1)
+  expect(called[3]).toBe(1)
+  expect(called[4]).toBe(1)
+  expect(called[5]).toBe(1)
+})
+
+test('forEach concurrency 1 base', async () => {
+  const arr = [...range(6)]
+  const called = {}
+  arr.forEach((v) => { called[v] = 0 })
+  await forEach(arr, async (x) => {
+    called[x] += 1
+  })
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(1)
+  expect(called[3]).toBe(1)
+  expect(called[4]).toBe(1)
+  expect(called[5]).toBe(1)
+})
+
+test('forEach concurrency 1 no async', async () => {
+  const arr = [...range(6)]
+  const called = {}
+  arr.forEach((v) => { called[v] = 0 })
+  await forEach(arr, (x) => {
+    called[x] += 1
+  })
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(1)
+  expect(called[2]).toBe(1)
+  expect(called[3]).toBe(1)
+  expect(called[4]).toBe(1)
+  expect(called[5]).toBe(1)
+})
+
+test('forEach concurrency 1 concurrency', async () => {
+  const arr = [...range(6)]
+  const called = {}
+  arr.forEach((v) => { called[v] = 0 })
+  const d = new Deferred()
+  const ds = arr.map(() => new Deferred())
+  const p = forEach(arr, async (x) => {
+    called[x] += 1
+    ds[x].resolve()
+    await d.promise
+  })
+  await ds[0].promise
+  expect(called[0]).toBe(1)
+  expect(called[1]).toBe(0)
+  expect(called[2]).toBe(0)
+  expect(called[3]).toBe(0)
+  expect(called[4]).toBe(0)
+  expect(called[5]).toBe(0)
   d.resolve()
   await p
   expect(called[0]).toBe(1)
