+++ src/modern-async.mjs
@@ -13,8 +13,6 @@ export { default as filterGenerator } from './filterGenerator.mjs'
 export { default as find } from './find.mjs'
 export { default as findIndex } from './findIndex.mjs'
 export { default as forEach } from './forEach.mjs'
-export { default as forEachLimit } from './forEachLimit.mjs'
-export { default as forEachSeries } from './forEachSeries.mjs'
 export { default as map } from './map.mjs'
 export { default as mapGenerator } from './mapGenerator.mjs'
 export { default as mapLimit } from './mapLimit.mjs'
